+++ src/data/paths.rs
@@ -1,16 +1,13 @@
-use std::path::{Path, PathBuf};
-
-use time::OffsetDateTime;
+use std::path::PathBuf;
 
 #[derive(Clone)]
 pub struct Paths {
     pub data_dir: PathBuf,
     pub daemon_dir: PathBuf,
     pub tasks_dir: PathBuf,
-    pub old_tasks_dir: PathBuf,
 
     pub tasks_file: PathBuf,
-    pub global_history_file: PathBuf,
+    pub history_file: PathBuf,
 
     pub daemon_socket_file: PathBuf,
     pub daemon_log_file: PathBuf,
@@ -22,50 +19,19 @@ impl Paths {
 
         Self {
             tasks_file: data_dir.join("tasks.json"),
-            global_history_file: data_dir.join("history.json"),
+            history_file: data_dir.join("history.json"),
 
             tasks_dir: data_dir.join("tasks"),
-            old_tasks_dir: data_dir.join("tasks.old"),
 
             daemon_socket_file: daemon_dir.join("daemon.sock"),
-            daemon_log_file: daemon_dir.join("output.log"),
+            daemon_log_file: daemon_dir.join("daemon.log"),
 
             daemon_dir,
             data_dir,
         }
     }
 
-    pub fn task_paths(&self, task_name: &str) -> TaskPaths {
-        TaskPaths {
-            task_dir: self.tasks_dir.join(task_name),
-        }
-    }
-
-    pub fn generate_old_task_dir_name(&self, task_name: &str) -> PathBuf {
-        self.old_tasks_dir.join(format!(
-            "{}-{}",
-            OffsetDateTime::now_local()
-                .expect("Failed to get current date/time")
-                .unix_timestamp(),
-            task_name
-        ))
-    }
-}
-
-pub struct TaskPaths {
-    task_dir: PathBuf,
-}
-
-impl TaskPaths {
-    pub fn dir(&self) -> &Path {
-        &self.task_dir
-    }
-
-    pub fn history_file(&self) -> PathBuf {
-        self.task_dir.join("history")
-    }
-
-    pub fn log_file(&self) -> PathBuf {
-        self.task_dir.join("output.log")
+    pub fn task_log_file(&self, task_name: &str) -> PathBuf {
+        self.tasks_dir.join(format!("{task_name}.log"))
     }
 }