+++ src/utils/save.rs
@@ -1,7 +1,5 @@
 use std::{
-    fs::{self, OpenOptions},
-    io::Read,
-    os::unix::prelude::FileExt,
+    fs::{self},
     path::{Path, PathBuf},
 };
 
@@ -9,7 +7,7 @@ use anyhow::{Context, Result};
 
 use crate::{
     history::{History, HistoryEntry},
-    paths::{Paths, TaskPaths},
+    paths::Paths,
     task::Tasks,
 };
 
@@ -36,10 +34,6 @@ pub fn ensure_data_dirs_exist(paths: &Paths) -> Result<()> {
         fs::create_dir(&paths.tasks_dir).context("Failed to create the tasks directory")?;
     }
 
-    if !paths.old_tasks_dir.exists() {
-        fs::create_dir(&paths.old_tasks_dir).context("Failed to create the old tasks directory")?;
-    }
-
     Ok(())
 }
 
@@ -63,14 +57,12 @@ pub fn write_tasks(paths: &Paths, tasks: &Tasks) -> Result<()> {
     fs::write(&paths.tasks_file, raw).context("Failed to write the tasks file")
 }
 
-pub fn read_history_file(task_paths: &TaskPaths) -> Result<Option<History>> {
-    let history_file = task_paths.history_file();
-
+pub fn read_history_file(history_file: &Path) -> Result<Option<History>> {
     if !history_file.is_file() {
         return Ok(None);
     }
 
-    let raw = fs::read_to_string(&history_file).context("Failed to read history file")?;
+    let raw = fs::read_to_string(history_file).context("Failed to read history file")?;
 
     let history = serde_json::from_str(&raw).context("Failed to parse the history file")?;
 
@@ -78,28 +70,17 @@ pub fn read_history_file(task_paths: &TaskPaths) -> Result<Option<History>> {
 }
 
 pub fn append_to_history(history_file: &Path, entry: HistoryEntry) -> Result<()> {
-    let mut file = OpenOptions::new()
-        .create(true)
-        .read(true)
-        .write(true)
-        .open(history_file)
-        .context("Failed to open the history file")?;
+    let mut history = if history_file.exists() {
+        let content = fs::read_to_string(history_file).context("Failed to read history file")?;
 
-    let mut content = String::new();
-    file.read_to_string(&mut content)
-        .context("Failed to read history file")?;
-
-    let mut history = if content.is_empty() {
-        History::empty()
-    }  else {    
         serde_json::from_str::<History>(&content).context("Failed to parse history file")?
+    } else {
+        History::empty()
     };
 
     history.append(entry);
 
-    let content = serde_json::to_string(&history).unwrap();
-
-    file.write_all_at(content.as_bytes(), 0)
+    fs::write(history_file, serde_json::to_string(&history).unwrap())
         .context("Failed to update history file")?;
 
     Ok(())